pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "aas-readable"
version = "0.4.0"
description = "Python API and CLI for converting Asset Administration Shell (AAS) and AASX data into lossless, deterministic agent and review representations"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
  {name = "Ethan Rozee"},
]
keywords = [
  "asset administration shell",
  "aas",
  "aasx",
  "digital twin",
  "industry 4.0",
  "llm",
  "agents",
  "graphrag",
  "semantic search",
  "manufacturing software",
  "json",
  "yaml",
  "markdown",
]
urls = {Homepage = "https://github.com/erozee1/AAS-Readable", Repository = "https://github.com/erozee1/AAS-Readable"}
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Information Technology",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Documentation",
  "Topic :: Scientific/Engineering :: Information Analysis",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Text Processing :: Markup",
  "Topic :: Scientific/Engineering",
]
dependencies = [
  "PyYAML>=6.0",
]

[project.optional-dependencies]
aasx = [
  "basyx-python-sdk>=1.2",
]

[project.scripts]
aas-readable = "aas_readable.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
